Convert muranoclient to 'requests'
Converts common/http.py to use requests, and adds unit tests. Most
of the code and tests were lifted from Heat (tests translated from
mox to mocks).

Because requests uses 'data' as its attribute for passing data to
requests, all uses of 'body' (for requests, not responses) have been
converted to use 'data'. This creates an incompatibility where 'body'
is explicitly passed to managers but these should be very rare. If
'body' is used, it will be converted to 'data' and a deprecation
warning issued.

Change-Id: I905bfb80642bfe19d7ebbdeb0fb8aaa92baefeb2
Closes-Bug: 1326804