Support multiple filenames for package commands
Before only a single package/bundle could be imported and deleted at a time.
This commit introduces support for multiple positional arguments in
bundle-import and package-import and package-delete commands.

Package and bundle dependencies are tracked in a single
dictionary during import. This means, that importing 2 packages,
that have common dependency-package would cause that package
to be imported only once.

Since it is highly complicated to specify --version for multiple
positional arguments with argparse this option is ignored, when used
with multiple packages.

Closes-Bug: #1441204
Partially-implements: blueprint multiple-positional-arguments
Change-Id: Id06c642d88f7d2a2726aec56e753c8db7dcc8ceb