Limit parameter is no longer ignored for packages
A 'packages.filter()' method was ignoring the 'limit' kwarg parameter by
overriding it with the value of 'page_size' parameter of the
DEFAULT_PAGE_SIZE constant if the 'page_size' was not set.

'page_size' is used by the dashboard, limit - by the engine, and right
now both of them have the same assumed semantics.

This change makes the client accept both of them (with priority for
'page_size'). This approach should be changed later, if we introduce
some client-side filtering: in this case 'limit' should define the
number of items requested by the client and the page_size should define
the number of items got from the API in each request. Right now this is
not the case, so these two parameters are quite interchangeable.

Change-Id: Ie4f4d1751bc5c12e46f230152a462c9b7a45ccd1
Partial-bug: #1445470