Only delete 'owned' packages for --exists-action update
Due to the fact that fqns are no longer unique update might currently
not work correctly, because it searches for package to delete by fqn.
To improve this situation this patch attempts to search for the package
in the current tenant and deletes it by id.
In case it is not found - raise the error, cause it most likely means
that the package is in another tennant and deleting it automatically is
dangerous.
This change relies on 'owned' parameter in api sorting, fixed
in https://review.openstack.org/#/c/177796/

Change-Id: Iee04ea65f5c4937ded4e259e3c0cb64189801da2
Closes-Bug: #1448135