Remove hash check during image upload
There were situations, when image was upload twice automatically,
because it was updated and it's hash changed.
But since these images have the same name and some of our applications
currently rely on image name, when constructing heat template it caused
problems in heat, as it could not resolve this situation,
while deploying template, generated by Murano.

Change-Id: Ib8f86f4934b2420d3c1c7e952dcbb47bcc9ab81c
Closes-Bug: #1443936