'to_dict' method added to PackageWrapper class
PackageWrapper objects, returned by the glare wrapper of the murano
client did not contain the 'to_dict' method which is needed for
carousel.

Appropriate method has been added. It does not need to return the full
dict representation of the object: just id, name and owner_id have to
be returned, as these are the only fields used by the carousel.

Change-Id: I2759ade3cfcbade27b459ba9b21b29e240f7696e
Closes-Bug: #1501405