CLI shell can now properly interact with Glance Artifacts
Since murano packages can now be stored in Glance Artifact Repository
the client needs to be instructed which service to query for them.
Appropriate argument was added to python client, however it was not
set from the shell for the CLI usage.

This patch adds appropriate command line argument to the shell script
and adds ability to instantiate artifacts_client to be passed to
muranoclient as an option.

(cherry picked from commit 84bc7eb)
Closes-bug: #1537010

Change-Id: I1c559f03f3d3d345697ec5ecbc95a16410500572