CLI shell can now properly interact with Glance Artifacts
Since murano packages can now be stored in Glance Artifact Repository
the client needs to be instructed which service to query for them.
Appropriate argument was added to python client, however it was not
set from the shell for the CLI usage.

This patch adds appropriate command line argument to the shell script
and adds ability to instantiate artifacts_client to be passed to
muranoclient as an option.

Change-Id: I2aabc2c98ee001dfe42dbc5c6d50cb279a5a10ce
Closes-bug: #1537010