Fix package delete on update when using Glare
Since Glare is unable to filter artifacts by "owned' flag, it should
not be passed as one of list query params.
Such a call is made when cli shell attempts to delete the conflicting
package during the package import process.

Another issue was that durint the package conflict CLI was searching
for confilicting package by its name only, ignoring version, thus it
was unable to delete proper package.

Change-Id: Idc7c7f552576f0c242e90d30cfc1a3f65c581e6d
Closes-Bug: #1539031