Fix reading of app template file when adding to env template
When running env-template-add-app subcommand via CLI, user is supposed to
pass "application template file" as a second argument. It seems that there
is no documentation for this template file format but the code shows that
this is a simple pass-through to the API which expects JSON (documented in
the API docs). However current CLI code reads this file as a set of plain
text lines and there's no sign of JSON parsing.

This commit changes this to load the file as JSON and pass further as python
object.

Change-Id: I53fe295424dcaf2ab4fa7e165c30a571d021ef21
Closes-Bug: #1545513