Use upper-constraints in tox test environments
This will force pip install to use the upper-constraints.txt specified
version of pip modules. When you don't do this, you are out on the
bleeding edge and become unstable everytime some python library in the
wide world changes in a way that you don't expect.

The releasenotes job has no constraints file, so override the install
command there.

Change-Id: I531606ce2ed981bddc618dac046ab081de2805eb
Partial-bug: #1598906