Make --limit argument in package-list command robust
If --limit argument was provided to command "murano package-list",
the result of the command was the same as if this argument wasn't
provided (in case of using murano db backend). The only difference
was that several requests to API were made instead of one (each
one requests <limit> packages at once).
This patch fixes limits number of packages in output of the command
and adds --marker argument to specify id of the package to start
request from.

Change-Id: Ia7647eebb7b22bfb47aacc93ca237915d69f8643
Closes-bug: #1605860