Populate tenant information in client
Python-muranoclient uses tenant id when talking to glare to avoid
importing the same package twice. However tenant information was not
available in the CLI and tenant id defaulted to None.
This led to a situation, where it was possible to
import the same package multiple times if it was not public.

This patch fixes the issue by acquiring package information from
keystoneauth or from user-supplied values in case no-auth is enforced.

Murano dashboard was not affected by this issue.

Change-Id: I72ad006c76945bf5707172ab128a89f1f6a7b112
Closes-Bug: #1624943