Fix Murano client to use V3 and MultiDomain Authentication
This fix will going to allow the usage Murano client with V3 Authentication
and multidomain environments.

Based on the latest template of RC where is used os_user_domain_name
instead of os_project_domain_id or os_user_domain_id.

if this variables doesn't exists will going to use the old method that was
working on older versions.

Change-Id: I420a6d825fe228279fb21bcdd5c172c9fc6369f5
Closes-Bug: #1600181