Ignore default values for deprecated security parameters
The command line parameters, ca-file, key-file, and cert-file,
are deprecated.  The problem is that the argument parser stores
the values for the deprecated parameters into the same attributes
used by Keystone.  As a result the murano dashboard is overriding
the default values of the Keystone attributes with its own default
values. This is wrong.

The change is to ignore these parameters if they are not specified
on the command line.  This will allow the Keystone client's method,
register_cli_options, to set them to the correct default values.
If the user specifies these parameters on the command line, then
we will honor them.

Change-Id: I08974f7fea0cd0cb2a906e26c9f197209c473b9b
Closes-bug: #1767144